feat(repository) : generic base repository to encapsulate common database CRUD
