+++ src/controllers/Interface/user/IUserController.ts
@@ -11,7 +11,6 @@ export interface IUserController {
   bookslot(req: Request, res: Response): Promise<void>
   reportAdvisor(req:Request, res:Response): Promise<Response>
   fetchSlotsByUser(req:Request, res:Response): Promise<Response>
-  getAdvisors(req: Request, res:Response): Promise<Response>
   createReview(req: Request, res:Response): Promise<Response>
   createGoal(req: Request, res:Response): Promise<Response>
   getGoalsById(req: Request, res:Response): Promise<Response>
