+++ src/repositories/Implementation/chatRepository.ts
@@ -9,7 +9,6 @@ import IUser from "../../entities/userEntities";
 import IAdvisor from "../../entities/advisorEntities";
 import advisorSchema from "../../models/advisorSchema";
 import chatSchema from "../../models/chatSchema";
-import notificationSchema, { INotification } from "../../models/notificationSchema";
 
 const isValidObjectId = (id: string): boolean => {
     return mongoose.Types.ObjectId.isValid(id);
@@ -209,36 +208,7 @@ export default class ChatRepository implements IChatRepository {
         }
     }
 
-    async createNotification(notification: Partial<INotification>): Promise<INotification> {
-        const newNotification = await notificationSchema.create(notification);
-        return newNotification
-      }
-    
-      async getNotificationsByUserId(receiverId: string): Promise<INotification[]> {
-        return await notificationSchema.find({ receiverId,isRead:false }).sort({ createdAt: -1 })
-      }
-    
-      async markAsRead(notificationId: string): Promise<INotification | null> {
-        return await notificationSchema.findByIdAndUpdate(
-          notificationId, 
-          { read: true }, 
-          { new: true }
-        );
-      }
-    
-      async markAllAsRead(receiverId: string): Promise<boolean> {
-        const result = await notificationSchema.updateMany(
-            { receiverId, isRead: false }, 
-            { $set: { isRead: true } }     
-        );
-    
-        return result.modifiedCount > 0;
-    }
-    
-      async deleteNotification(notificationId: string): Promise<boolean> {
-        const result = await notificationSchema.findByIdAndDelete(notificationId);
-        return !!result;
-      }
+
 
     // async saveNotification(senderId: string, receiverId: string, message: string, link: string, type: 'message' | 'payment' | 'job'): Promise<INotification> {
     //     return await Notification.create({ senderId, receiverId, message, link, type })
