+++ src/repositories/Interface/IChatRepository.ts
@@ -3,7 +3,6 @@ import { IFriendsLists } from "../../entities/friendsEntities"
 import IUser from "../../entities/userEntities";
 import { IChat } from "../../models/chatSchema";
 import { IMessage } from "../../models/messageSchema";
-import { INotification } from "../../models/notificationSchema";
 
 
 
@@ -16,11 +15,7 @@ export interface IChatRepository {
     createChat(chatData: Partial<IChat>): Promise<IChat>
     getUserChats(id: string): Promise<IChat[]>
     getAllChats(): Promise<IChat[]>
-    createNotification(notification: Partial<INotification>): Promise<INotification>;
-    getNotificationsByUserId(userId: string): Promise<INotification[]>;
-    markAsRead(notificationId: string): Promise<INotification | null>;
-    markAllAsRead(userId: string): Promise<boolean>;
-    deleteNotification(notificationId: string): Promise<boolean>;
+
 
 
     //not below methods . so just keep it like that
