+++ src/repositories/Interface/IReviewRepository.ts
@@ -1,0 +1,6 @@
+import { IReview } from "../../models/reviewSchema";
+
+export interface IReviewRepository{
+    createReview(advisorId: string, userId: string, rating: number, review: string): Promise<IReview>
+
+}
