+++ src/routes/user/userRouter.ts
@@ -36,7 +36,6 @@ router.post('/paymentInitiate',AuthMiddleware.authorizeUser,(req,res)=>paymentCo
 router.post('/confirmPayment',AuthMiddleware.authorizeUser,(req,res)=>paymentController.confirmPayment(req,res))
 router.post('/reportAdvisor/:slotId',AuthMiddleware.authorizeUser,(req,res)=>userController.reportAdvisor(req,res))
 router.get('/fetchSlotsByUser/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.fetchSlotsByUser(req,res))
-router.get('/getAdvisors',AuthMiddleware.authorizeUser,(req,res)=>userController.getAdvisors(req,res))
 router.post('/createReview',AuthMiddleware.authorizeUser,(req,res)=>userController.createReview(req,res))
 router.post('/createGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.createGoal(req,res))
 router.get('/getGoals/:userId',AuthMiddleware.authorizeUser,(req,res)=>userController.getGoalsById(req,res))
